+++ driver/web/docs/gen/gen_types.go
@@ -580,6 +580,12 @@ type ServiceScope struct {
 	Scope       string  `json:"scope"`
 }
 
+// AdminReqAddApplicationAccountGroup defines model for _admin_req_add-application_account_group.
+type AdminReqAddApplicationAccountGroup struct {
+	AccountIds []string `json:"account_ids"`
+	GroupId    string   `json:"group_id"`
+}
+
 // AdminReqCreateApplicationGroup defines model for _admin_req_create-application_group.
 type AdminReqCreateApplicationGroup struct {
 	Name        string    `json:"name"`
@@ -1178,6 +1184,9 @@ type GetAdminApplicationAccountsParams struct {
 	AuthTypeIdentifier *string `json:"auth-type-identifier,omitempty"`
 }
 
+// PutAdminApplicationGroupAccountsAddJSONBody defines parameters for PutAdminApplicationGroupAccountsAdd.
+type PutAdminApplicationGroupAccountsAddJSONBody AdminReqAddApplicationAccountGroup
+
 // PostAdminApplicationGroupsJSONBody defines parameters for PostAdminApplicationGroups.
 type PostAdminApplicationGroupsJSONBody AdminReqCreateApplicationGroup
 
@@ -1576,6 +1585,9 @@ type PutAdminApplicationAccountPermissionsGrantJSONRequestBody PutAdminApplicati
 
 // PutAdminApplicationAccountRolesGrantJSONRequestBody defines body for PutAdminApplicationAccountRolesGrant for application/json ContentType.
 type PutAdminApplicationAccountRolesGrantJSONRequestBody PutAdminApplicationAccountRolesGrantJSONBody
+
+// PutAdminApplicationGroupAccountsAddJSONRequestBody defines body for PutAdminApplicationGroupAccountsAdd for application/json ContentType.
+type PutAdminApplicationGroupAccountsAddJSONRequestBody PutAdminApplicationGroupAccountsAddJSONBody
 
 // PostAdminApplicationGroupsJSONRequestBody defines body for PostAdminApplicationGroups for application/json ContentType.
 type PostAdminApplicationGroupsJSONRequestBody PostAdminApplicationGroupsJSONBody