+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Service accounts [#306](https://github.com/rokwire/core-building-block/issues/306)
+- Expose add accounts to a group Admin API [#384](https://github.com/rokwire/core-building-block/issues/384)
 - Handle external ID management [#364](https://github.com/rokwire/core-building-block/issues/364)
 
 ## [1.21.1] - 2022-03-17